+++ lib/solidus_subscriptions/engine.rb
@@ -31,8 +31,8 @@ module SolidusSubscriptions
     end
 
     initializer 'solidus_subscriptions.register_promotion_rules', after: 'spree.promo.register.promotion.rules' do |app|
-      app.config.spree.promotions.rules << 'SolidusSubscriptions::SubscriptionCreationOrderPromotionRule'
-      app.config.spree.promotions.rules << 'SolidusSubscriptions::SubscriptionInstallmentOrderPromotionRule'
+      app.config.spree.promotions.rules << 'SolidusSubscriptions::Promotion::Rules::SubscriptionCreationOrder'
+      app.config.spree.promotions.rules << 'SolidusSubscriptions::Promotion::Rules::SubscriptionInstallmentOrder'
     end
 
     initializer 'solidus_subscriptions.configure_backend' do
