+++ spec/lib/solidus_subscriptions/promotion/rules/subscription_creation_order_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe SolidusSubscriptions::SubscriptionCreationOrderPromotionRule do
+RSpec.describe SolidusSubscriptions::Promotion::Rules::SubscriptionCreationOrder do
   let(:rule) { described_class.new }
 
   describe '#applicable' do
