+++ spec/lib/solidus_subscriptions/promotion/rules/subscription_installment_order_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe SolidusSubscriptions::SubscriptionInstallmentOrderPromotionRule do
+RSpec.describe SolidusSubscriptions::Promotion::Rules::SubscriptionInstallmentOrder do
   let(:rule) { described_class.new }
 
   describe '#applicable' do
